moved dtos and refrenced player models
