+++ src/features/battle/dtos/requests/is-ready-request.ts
