+++ src/features/battle/dtos/responses/battle-result-response.ts
